Extract util for getting slots with updated recipient
