+++ src/handlers/expenditures/expenditureRecipientSet.ts
@@ -1,6 +1,5 @@
 import { mutate } from '~amplifyClient';
 import {
-  ExpenditureSlot,
   UpdateExpenditureDocument,
   UpdateExpenditureMutation,
   UpdateExpenditureMutationVariables,
@@ -8,7 +7,7 @@ import {
 import { ContractEvent } from '~types';
 import { getExpenditureDatabaseId, output, toNumber, verbose } from '~utils';
 
-import { getExpenditureFromDB } from './helpers';
+import { getExpenditureFromDB, getSlotsWithUpdatedRecipient } from './helpers';
 
 export default async (event: ContractEvent): Promise<void> => {
   const { contractAddress: colonyAddress } = event;
@@ -29,18 +28,11 @@ export default async (event: ContractEvent): Promise<void> => {
     return;
   }
 
-  const existingSlot = expenditure.slots.find(
-    (slot) => slot.id === convertedSlot,
-  );
-  const updatedSlot: ExpenditureSlot = {
-    ...existingSlot,
-    id: convertedSlot,
+  const updatedSlots = getSlotsWithUpdatedRecipient(
+    expenditure,
+    convertedSlot,
     recipientAddress,
-  };
-  const updatedSlots = [
-    ...expenditure.slots.filter((slot) => slot.id !== convertedSlot),
-    updatedSlot,
-  ];
+  );
 
   verbose(
     `Recipient set for expenditure with ID ${convertedExpenditureId} in colony ${colonyAddress}`,
