+++ src/handlers/expenditures/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from './getExpenditure';
 export * from './updateCancelledExpenditure';
+export * from './slots';
