+++ src/handlers/expenditures/helpers/slots.ts
@@ -1,0 +1,24 @@
+import { ExpenditureFragment, ExpenditureSlot } from '~graphql';
+
+/**
+ * Helper function returning expenditure slots with updated recipient
+ * of a given slot
+ */
+export const getSlotsWithUpdatedRecipient = (
+  expenditure: ExpenditureFragment,
+  slotId: number,
+  recipientAddress: string,
+): ExpenditureSlot[] => {
+  const existingSlot = expenditure.slots.find((slot) => slot.id === slotId);
+  const updatedSlot: ExpenditureSlot = {
+    ...existingSlot,
+    id: slotId,
+    recipientAddress,
+  };
+  const updatedSlots = [
+    ...expenditure.slots.filter((slot) => slot.id !== slotId),
+    updatedSlot,
+  ];
+
+  return updatedSlots;
+};
